# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BarberShop API",
        "description": "Appointment Scheduling System for Barbershops.",
        "contact": {
            "url": "https://github.com/matheusmrqs4"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/barber-shop/profile": {
            "get": {
                "tags": [
                    "BarberShop Profile"
                ],
                "summary": "Get BarberShop Profile",
                "operationId": "profile",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "barber_shop": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Error fetching profile"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/barber-shop/update-profile": {
            "put": {
                "tags": [
                    "BarberShop Profile"
                ],
                "summary": "Update BarberShop Profile",
                "operationId": "updateProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "BarberShop name",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "BarberShop phone",
                                        "type": "string"
                                    },
                                    "address1": {
                                        "description": "BarberShop Address line 1",
                                        "type": "string"
                                    },
                                    "address2": {
                                        "description": "BarberShop Address line 2",
                                        "type": "string"
                                    },
                                    "address3": {
                                        "description": "BarberShop Address line 3",
                                        "type": "string"
                                    },
                                    "bio": {
                                        "description": "BarberShop biography",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                },
                                                "barber_shop": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Error updating profile"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}